requirements.txt
aioflask==0.4.0
asgiref==3.5.2
click==8.1.3
connect-mongodb==0.0.1
dnspython==2.2.1
Flask==2.1.2
greenlet==1.1.2
greenletio==0.9.0
gunicorn==20.0.4
h11==0.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
uvicorn==0.17.6
Werkzeug==2.1.2
